avoid indexing browse and search pages
Search pages and browse/search results pages are 'light on content',
and might be considered 'low-quality' by search engines.

By explicitly tagging these pages as 'not to be indexed', they are
not counted when determining the 'overall site quality', which should
improve the overall site quality (and thus ranking) of Omeka-based sites.